Update max element in an array.cpp

+++ Beginner/max element in an array.cpp
@@ -2,9 +2,9 @@
 using namespace std;
 
 int main() {
-	int n,i,j,temp;
+	long int n,i,j,temp;
 	cin>>n;
-	int a[n];
+	long int a[n];
 	for(i=0;i<n;i++)
 	{
 	    cin>>a[i];
